Move ConnectionCacheStats to new module
