+++ quic-client/src/lib.rs
@@ -18,7 +18,7 @@ use {
         tls_certificates::new_self_signed_tls_certificate_chain,
     },
     solana_tpu_client::{
-        connection_cache::ConnectionCacheStats,
+        connection_cache_stats::ConnectionCacheStats,
         tpu_connection_cache::{BaseTpuConnection, ConnectionPool, ConnectionPoolError},
     },
     std::{
